# Remove use of functors for sparse matrices and vectors