--- a/base/sparse.jl
+++ b/base/sparse.jl
@@ -2,7 +2,6 @@
 
 module SparseArrays
 
-using Base: Func, AddFun, OrFun, ConjFun, IdFun
 using Base.Sort: Forward
 using Base.LinAlg: AbstractTriangular, PosDefException