One step towards https://github.com/JuliaLang/julia/issues/15692. cc @timholy