@KristofferC 2016-03-30
Did some sport benchmarks:

### triu!

```jl
# Master
julia> @time triu!(B);
  0.039143 seconds (4 allocations: 160 bytes)


# PR
julia> @time triu!(B);
  0.039338 seconds (5 allocations: 176 bytes)
```

### sparse
```jl
#Master
2.532434 seconds (27 allocations: 322.724 MB, 3.15% gc time)

#PR
2.347069 seconds (27 allocations: 322.724 MB, 3.80% gc time)
```

Enough for me...