@KristofferC 2016-03-30
I only ran the sparse tests locally but this seems to break something with a type promotion test for broadcasting in `test/arrayops.jl`