@KristofferC 2016-03-30
Ok, reverting back to 

```jl
(.*)(A::AbstractArray, B::AbstractArray) = broadcast_zpreserving(MulFun(), A, B)
```

should fix it. That is a pretty odd place to have a method definition for `.*` between all `AbstractArray`s...