@timholy 2016-03-30
If you're eager to eliminate this one too, you could presumably define a `broadcast_zpreserving(f::Function, ...)` and then have deprecated fallbacks like `broadcast(::MulFun, ...) = broadcast_zpreserving(*, ...)`.