@KristofferC 2016-03-30
Not sure I am that eager :P. Although, talking about deprecations.. does all of this need deprecations since some people might have written code that uses functors to call for example `sparse`?