@timholy 2016-03-30
If exported functions allow a functor input (I haven't checked), then presumably it needs a deprecation. Non-exported functions are at-your-own risk, so don't need one.

In terms of just deprecations, an easier path might be to deprecate just `Base.MulFun()` itself, and have it generate `*`. Then you don't need to write any deprecations for any functions that accept functor inputs. So one option is to let this PR sit until some nice person helps out with the overall task.