@KristofferC 2016-03-30
AppVeyor timed out.

Sparse used to allow `Union{Func, Function}` as the `combine` function. Letting this sit is ok by me.